+++ R/crane-package.R
@@ -1,5 +1,6 @@
 #' @keywords internal
 #' @import rlang
+#' @import glue glue
 #' @importFrom dplyr across starts_with ends_with contains matches num_range
 #'                   all_of any_of everything last_col where
 "_PACKAGE"
@@ -8,9 +9,9 @@
 ## usethis namespace: end
 NULL
 
-
-# silence notes about unused pkgs
-silenced <- function() {
-  cards::ard_continuous
-  cardx::ard_continuous
+# using pkgs to silence NOTE
+.silence <- function() {
+  # broom is used in tbl_survfit_quantiles() via cardx::ard_survfit()
+  broom::tidy
+  survival::Surv
 }